[FEAT] 이미지 확인 서비스 추가 및 기존 로직 리팩터링
src/main/java/com/teambind/image_server/service/ImageConfirmService.java: 이미지 확인/프로필 이미지 삭제 처리 서비스 구현
src/main/java/com/teambind/image_server/exception/ErrorCode.java: 기존 오류 ENUM에 IMAGE_ALREADY_CONFIRMED 코드 추가
src/main/java/com/teambind/image_server/repository/ImageRepository.java: 프로필 이미지 삭제를 위한 커스텀 메서드 추가
src/main/java/com/teambind/image_server/util/helper/SequenceHelper.java: SequenceHelper에 Repository 빈 추가
src/test/java/com/teambind/image_server/util/convertor/WebpConvertorTest.java: WebP 변환 관련 테스트 클래스 제거
src/main/java/com/teambind/image_server/service/ImageUploadService.java: 이미지 업로드 서비스 및 관련 로직 제거

참고: 기존 불필요한 테스트 및 업로드 서비스 클래스는 삭제 처리.

+++ src/main/java/com/teambind/image_server/service/ImageSaveService.java
@@ -80,7 +80,7 @@ public class ImageSaveService {
                 .convertedFormat(extensionMap.get("WEBP"))
                 .originFormat(extensionMap.get(extensionParser.extensionParse(fileName).toUpperCase()))
                 .originSize(file.getSize())
-                .convertedSize(webpBytes.length)
+                .convertedSize((long) webpBytes.length)
                 .storageLocation(storedPath)
                 .build();
